requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
scipy
tensorboardX